+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/note/NoteNodeResponseService.java
@@ -11,6 +11,8 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.Service;
 
+import java.util.List;
+
 /**
  * 笔记节点响应服务。
  *
@@ -32,6 +34,9 @@ public interface NoteNodeResponseService extends Service {
 
     PagedData<NoteNode> childForParent(LongIdKey parentKey, PagingInfo pagingInfo) throws ServiceException;
 
+    PagedData<NoteNode> childForNoteBookNameLike(LongIdKey noteBookKey, String pattern, PagingInfo pagingInfo)
+            throws ServiceException;
+
     DispNoteNode getDisp(LongIdKey key, StringIdKey inspectAccountKey) throws ServiceException;
 
     PagedData<DispNoteNode> allDisp(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException;
@@ -48,10 +53,23 @@ public interface NoteNodeResponseService extends Service {
             StringIdKey accountKey, LongIdKey parentKey, PagingInfo pagingInfo
     ) throws ServiceException;
 
+    PagedData<DispNoteNode> childForNoteBookNameLikeDisp(
+            StringIdKey accountKey, LongIdKey noteBookKey, String pattern, PagingInfo pagingInfo
+    ) throws ServiceException;
+
     LongIdKey createNoteNode(StringIdKey userKey, NoteNodeCreateInfo noteNodeCreateInfo)
             throws ServiceException;
 
     void updateNoteNode(StringIdKey userKey, NoteNodeUpdateInfo noteNodeUpdateInfo) throws ServiceException;
 
     void removeNoteNode(StringIdKey userKey, LongIdKey noteNodeKey) throws ServiceException;
+
+    /**
+     * 获取指定的笔记节点从根节点到该节点所属父节点的路径。
+     *
+     * @param key 指定的笔记节点的主键。
+     * @return 指定的笔记节点从根节点到该节点所属父节点的路径。
+     * @throws ServiceException 服务异常。
+     */
+    List<LongIdKey> pathFromRoot(LongIdKey key) throws ServiceException;
 }